Move OtaHandler to src OTA only valid for esp32 based monitor
